tests/wpt/web-platform-tests/referrer-policy/css-integration/image/presentation-attribute.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - image from presentation attribute</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that resources from presentation attributes are loaded with
    the referrer and referrer policy from the document.</p>

    <script>
      var css_test = async_test("Image from presentation attributes.");
      var id = token();
      var css_url = location.protocol + "//www1." + location.hostname + ":" + location.port + "/referrer-policy/generic/subresource/image.py" + "?id=" + id;
      var img_url = css_url + "&report-headers";

      document.body.background = css_url;
      css_test.step_timeout(
          queryXhr.bind(this, img_url,
                        function(message) {
                          css_test.step(function() {
                            assert_own_property(message, "headers");
                            assert_own_property(message, "referrer");
                            assert_equals(message.referrer, location.origin + "/");
                          });
                          css_test.done();
                        }),
          1000);
      </script>

    <div id="log"></div>
  </body>
</html>
